fast-api-template/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
h11==0.14.0
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pip==24.2
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
